Create add_javascript_to_limit_cart_quantity.php
First, you'll need to add a JavaScript code snippet to your theme's functions.php file or a separate JavaScript file. This code will restrict the maximum quantity of products that can be added to the cart based on the product's category.

This code enqueues a custom JavaScript file (limit-cart-quantity.js) on the product category pages specified by the category IDs in the $limited_categories array. The wp_localize_script function is used to pass the maximum allowed quantity (max_quantity) to the JavaScript file.